Use text/template to render variable replacement
This is a new implementation of ReplaceString which uses text/template for heavy lifting.

What it does is first the input is rewritten to tags that are going to be identified by text/emplate, then the rewritten template is passed to text/template for rendering.

The special tags are `<<` and `>>` so,  `{host}` becomes `<< replace "host">>` where replace is the template function with signature `func(string)string`

This will work well with json, the fact here is we can later change special tags to be unique to any particular text format that way the function can be used on xml etc.

The replaced variables are handled by the end user meaning not only can json objects be adressed but aslo arbitrary logic can be handled since templating is separaed from replacement